First Commit - ST
The first commit.

This description aims to summarise everything that has been changed since the last commit.

If anything is removed or added, it needs to be clear here.

Good practice is to comment out anything you have removed, so the team can verify that it needs to be removed.