Added sumToArray Function
Added the sumToArray function. This takes in two integers and an array.

The two integers are added together and appended to the array.

This is done by creating a new array one larger than the input array and appending the total to the end. The new array is then returned (with the intention of replacing the old array).

Also added some debug code in the main function to test that it functions as intended and to test other functions.